--- bankAccount-service/src/main/java/edu/miu/cs590/bankAccountservice/BARepository.java
@@ -1,9 +1,0 @@
-package edu.miu.cs590.bankAccountservice;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BARepository extends JpaRepository<Transactions, Integer> {
-
-    public Transactions findByCardNumber(String cardNumber);
-
-}
