--- bankAccount-service/src/main/java/edu/miu/cs590/bankAccountservice/BAService.java
@@ -1,8 +1,0 @@
-package edu.miu.cs590.bankAccountservice;
-
-public interface BAService {
-
-    public String processPayment(String username, double amount);
-    public String confirmPayment();
-
-}
