--- bankAccount-service/src/main/java/edu/miu/cs590/bankAccountservice/Transactions.java
@@ -1,31 +1,0 @@
-package edu.miu.cs590.bankAccountservice;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.transaction.Transactional;
-
-@Data
-@Transactional
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class Transactions {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
-    private String cardHolderName;
-    private String cardNumber;
-    private int CCV;
-    private double creditedAmount;
-    private String expiryDate;
-
-
-}
