+++ bankAccount-service/src/main/java/edu/miu/cs590/bankAccountservice/dao/PaymentRequest.java
@@ -1,0 +1,15 @@
+package edu.miu.cs590.bankAccountservice.dao;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class PaymentRequest implements java.io.Serializable {
+
+    private int orderId;
+    private String payment_type;
+    private String username;
+    private double payment_amount;
+
+}
