+++ bankAccount-service/src/main/java/edu/miu/cs590/bankAccountservice/service/BankService.java
@@ -1,0 +1,8 @@
+package edu.miu.cs590.bankAccountservice.service;
+
+public interface BankService {
+
+    public String processPayment(String username, double amount);
+    public String confirmPayment();
+
+}
