--- creditCard-service/src/main/java/edu/miu/cs590/creditCardservice/CCRepository.java
@@ -1,9 +1,0 @@
-package edu.miu.cs590.creditCardservice;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CCRepository extends JpaRepository<Transactions, Integer> {
-
-    public Transactions findByCardNumber(String cardNumber);
-
-}
