--- creditCard-service/src/main/java/edu/miu/cs590/creditCardservice/CCService.java
@@ -1,8 +1,0 @@
-package edu.miu.cs590.creditCardservice;
-
-public interface CCService {
-
-    public String processPayment(String username, double amount);
-    public String confirmPayment();
-
-}
