+++ creditCard-service/src/main/java/edu/miu/cs590/creditCardservice/Repository/CCRepository.java
@@ -1,0 +1,10 @@
+package edu.miu.cs590.creditCardservice.Repository;
+
+import edu.miu.cs590.creditCardservice.Transactions;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CCRepository extends JpaRepository<Transactions, Integer> {
+
+    public Transactions findByCardNumber(String cardNumber);
+
+}
