+++ creditCard-service/src/main/java/edu/miu/cs590/creditCardservice/service/CCService.java
@@ -1,0 +1,8 @@
+package edu.miu.cs590.creditCardservice.service;
+
+public interface CCService {
+
+    public String processPayment(String username, double amount);
+    public String confirmPayment();
+
+}
