--- payPal-service/src/main/java/edu/miu/cs590/payPalservice/PPRepository.java
@@ -1,9 +1,0 @@
-package edu.miu.cs590.payPalservice;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PPRepository extends JpaRepository<Transactions, Integer> {
-
-    public Transactions findByCardNumber(String cardNumber);
-
-}
