--- payPal-service/src/main/java/edu/miu/cs590/payPalservice/PPService.java
@@ -1,8 +1,0 @@
-package edu.miu.cs590.payPalservice;
-
-public interface PPService {
-
-    public String processPayment(String username, double amount);
-    public String confirmPayment();
-
-}
