+++ payPal-service/src/main/java/edu/miu/cs590/payPalservice/Service/PPService.java
@@ -1,0 +1,8 @@
+package edu.miu.cs590.payPalservice.Service;
+
+public interface PPService {
+
+    public String processPayment(String username, double amount);
+    public String confirmPayment();
+
+}
