+++ payPal-service/src/main/java/edu/miu/cs590/payPalservice/repository/Repository.java
@@ -1,0 +1,10 @@
+package edu.miu.cs590.payPalservice.repository;
+
+import edu.miu.cs590.payPalservice.entity.Transactions;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface Repository extends JpaRepository<Transactions, Integer> {
+
+    public Transactions findByCardNumber(String cardNumber);
+
+}
