--- payment-service/src/main/java/edu/miu/cs590/paymentservice/PaymentAPI.java
@@ -1,39 +1,0 @@
-package edu.miu.cs590.paymentservice;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("payment")
-public class PaymentAPI {
-
-    @Autowired
-    private PaymentService paymentService;
-
-    // a post function to create a payment
-    @PostMapping
-    public void createPayment(@RequestBody PaymentRequest paymentRequest) {
-        paymentService.createPayment(paymentRequest);
-
-    }
-
-
-    @GetMapping("/{id}")
-    public String getPaymentById(@PathVariable int id) {
-
-        return paymentService.getStatusByOrderId(id);
-    }
-
-    @GetMapping
-    public List<Payment> getAllPayments() {
-
-        return paymentService.getAllPayments();
-    }
-
-
-
-
-
-}
