--- payment-service/src/main/java/edu/miu/cs590/paymentservice/PaymentRepository.java
@@ -1,12 +1,0 @@
-package edu.miu.cs590.paymentservice;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PaymentRepository extends JpaRepository<Payment, Integer> {
-
-
-    Payment getPaymentsByOrderId(int orderId);
-
-}
-
-
