+++ payment-service/src/main/java/edu/miu/cs590/paymentservice/config/Config.java
@@ -1,0 +1,14 @@
+package edu.miu.cs590.paymentservice.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class Config {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
