+++ payment-service/src/main/java/edu/miu/cs590/paymentservice/dao/PaymentRequest.java
@@ -1,0 +1,15 @@
+package edu.miu.cs590.paymentservice.dao;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class PaymentRequest implements java.io.Serializable {
+
+    private Long orderId;
+    private String defaultPaymentMethod;
+    private Long userId;
+    private double price;
+
+}
