+++ payment-service/src/main/java/edu/miu/cs590/paymentservice/entity/Payment.java
@@ -1,0 +1,47 @@
+package edu.miu.cs590.paymentservice.entity;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.transaction.annotation.Transactional;
+
+import javax.persistence.Entity;
+import java.time.LocalDate;
+import javax.persistence.*;
+
+@Entity
+@Builder
+@Transactional
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Payment {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(name = "payment_type")
+    private String preferredPaymentMethod;
+
+    @Column(name = "order_id")
+    private Long orderId;
+
+    @Column(name = "user_id")
+    private Long userId;
+
+    @Column(name = "payment_date")
+    private LocalDate paymentDate;
+
+    @Column(name = "payment_amount")
+    private Double paymentAmount;
+
+    @Column(name = "payment_status")
+    private String paymentStatus;
+
+
+
+
+}
