+++ payment-service/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/cs590PaymentService
 spring.datasource.username=root
-spring.datasource.password=speedo2942
+spring.datasource.password=password
 
 spring.jpa.hibernate.ddl-auto=create-drop
